+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The gtk4 branch is the new development branch.
 - examples/gtkhello.f90: new version using GtkApplication.
 - Most examples are now using GtkApplication.
 - cfwrapper.py is now systematically launching extract_events.pl. Backported to gtk3 branch.
+- cfwrapper.py is now updating the codemeta.json file (dateModified field). Backported to gtk3 branch.
 - VERSIONS : the name of the distribution used to generate gtk-fortran is now automatically found. Backported to the gtk3 branch.
 
 ### Removed